+++ mobile/app/(tabs)/news.tsx
@@ -1,9 +1,20 @@
-import { Text, View } from 'tamagui';
+import PageTitle from "@/src/components/PageTitle";
+import PostCard from "@/src/components/PostsScreen/PostCard";
+import { api } from "@/utils/api";
+import { ScrollView, YStack } from "tamagui";
 
 export default function News() {
+  const posts = api.post.getPosts.useQuery();
+
   return (
-    <View flex={1} jc={'center'} ai={'center'}>
-      <Text>Notícias</Text>
-    </View>
+    <ScrollView>
+      <YStack margin={16} style={{ flex: 1, gap: 16 }}>
+        <PageTitle title="Notícias"></PageTitle>
+        {posts.data &&
+          posts.data.map((post) => {
+            return <PostCard key={post.id} post={post}></PostCard>;
+          })}
+      </YStack>
+    </ScrollView>
   );
 }
